src/main/java/org/example/socialmedia/service/FollowService.java
package org.example.socialmedia.service;

import org.example.socialmedia.entity.Follow;
import org.example.socialmedia.repository.FollowRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class FollowService {
    @Autowired
    private FollowRepository followRepository;

    public List<Follow> getAllFollows() {
        return followRepository.findAll();
    }

    public Follow getFollowById(Long id) {
        return followRepository.findById(id).orElse(null);
    }

    public Follow createFollow(Follow follow) {
        return followRepository.save(follow);
    }

    public void deleteFollow(Long id) {
        followRepository.deleteById(id);
    }
}